requirements.txt
SpeechRecognition
pyaudio
gTTS
playsound==1.2.2
arabic-reshaper
python-bidi
